backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
boto3==1.34.69
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
